+++ src/TestLibrary/StringUtils.cs
@@ -1,0 +1,7 @@
+﻿namespace TestLibrary
+{
+    public static class StringUtils
+    {
+        public static string TrimmedString(string value) => value.Trim('\r', '\n');
+    }
+}